% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{count_sentences}
\alias{count_sentences}
\title{Count sentences}
\usage{
count_sentences(text)
}
\arguments{
\item{text}{Character}
}
\value{
IntegerVector.
}
\description{
Count sentences
}
\examples{
count_sentences("Hello world! My name is Mike. Why so serious?")
}
