% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_syllables.R
\name{count_syllables}
\alias{count_syllables}
\title{Counts the number of syllables in a word}
\usage{
count_syllables(word)
}
\arguments{
\item{word}{Character.}
}
\value{
Character.
}
\description{
Counts the number of syllables in a word
}
