% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_words.R
\name{count_words}
\alias{count_words}
\title{Counts the number of words in text.}
\usage{
count_words(text)
}
\arguments{
\item{text}{Character.}
}
\value{

}
\description{
Counts the number of words in text.
}
