% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complex_word.R
\name{filter_complex_words}
\alias{filter_complex_words}
\title{Examines if a word is complex, that is, when it
contains three or more syllables.}
\usage{
filter_complex_words(words)
}
\arguments{
\item{words}{Character Vector.}
}
\value{

}
\description{
Examines if a word is complex, that is, when it
contains three or more syllables.
}
