% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{get_sentences}
\alias{get_sentences}
\title{Extract sentences from text.}
\usage{
get_sentences(text)
}
\arguments{
\item{text}{Character of length one.}
}
\value{
CharacterVector.
}
\description{
Extract sentences from text.
}
\examples{
get_sentences("Hello world! My name is Mike. Why so serious?")
}
