% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gunning_fog_idx.R
\name{gunning_fog}
\alias{gunning_fog}
\title{Quantify the complexity of text with the Gunning Fox Index.}
\usage{
gunning_fog(text)
}
\arguments{
\item{text}{Character.}
}
\value{

}
\description{
Quantify the complexity of text with the Gunning Fox Index.
}
