% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_complex_word.R
\name{is_complex_word}
\alias{is_complex_word}
\title{Examines if a word is complex, where complexity is
defined by the number of syllables in a word (n).}
\usage{
is_complex_word(word, n = 3)
}
\arguments{
\item{word}{Character.}

\item{n}{int. Number of syllables in a word before a word counts as complex.}
}
\value{
Logical
}
\description{
Examines if a word is complex, where complexity is
defined by the number of syllables in a word (n).
}
