% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/most_similar.R
\name{most_similar}
\alias{most_similar}
\title{Find the most similar string in a target set of strings.}
\usage{
most_similar(strings, targets, method = c("dl", "lcs", "lv", "jw", "jaro"))
}
\arguments{
\item{strings}{String Vector. Strings that you want to compare with a target.}

\item{targets}{String Vector. Set of target strings.}

\item{method}{Edit distance method. Either "dl", "lcs", "lv", "jw", or "jaro".}
}
\value{
List
}
\description{
Find the most similar string in a target set of strings.
}
\examples{
most_similar(c("some", "same"), c("same", "samex", "some"), method = "lv")
}
