% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_similar_jaro}
\alias{most_similar_jaro}
\title{Calculate for each string the Jaro Distance given a set of target strings.}
\usage{
most_similar_jaro(strings, targets)
}
\arguments{
\item{strings}{String Vector.}

\item{targets}{String Vector.}
}
\description{
Calculate for each string the Jaro Distance given a set of target strings.
}
