% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{most_similar_levenstein}
\alias{most_similar_levenstein}
\title{Calculate for each string the normalized Levenstein Distance given a set of target strings.}
\usage{
most_similar_levenstein(strings, targets)
}
\arguments{
\item{strings}{String Vector.}

\item{targets}{String Vector.}
}
\description{
Calculate for each string the normalized Levenstein Distance given a set of target strings.
}
