% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{count_sentences}
\alias{count_sentences}
\title{Count the number of sentences in text.}
\usage{
count_sentences(text)
}
\arguments{
\item{text}{CharacterVector.}
}
\value{

}
\description{
Count the number of sentences in text.
}
