% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentences.R
\name{str_extract}
\alias{str_extract}
\title{Extract data from text given a regular expression.}
\usage{
str_extract(string, pattern)
}
\arguments{
\item{string}{Character.}

\item{pattern}{A regular expression.}
}
\value{

}
\description{
Extract data from text given a regular expression.
}
